refactor(identifiers): split government/detection.rs into per-country submodules
Slice of the #193 code-health umbrella — file-length category. The single
file primitives/identifiers/government/detection.rs grew to 2060 LOC.
Split into a detection/ directory of 17 files mirroring the existing
sibling validation/ per-country layout:

  detection/mod.rs          (top-level dispatchers + re-exports, 347 LOC)
  detection/helpers.rs      (shared regex/dedup helpers, 97 LOC)
  detection/ssn.rs          (SSN + SSA filtering, 276 LOC)
  detection/tax_id.rs       (EIN/ITIN/TIN, 203 LOC)
  detection/national_id.rs  (generic + UK NI, 250 LOC)
  detection/india.rs        (Aadhaar, PAN, GSTIN, …, 222 LOC)
  detection/europe.rs       (Spain, Italy, Finland, Poland, 183 LOC)
  + 10 smaller country/domain files (40-110 LOC each)

Pure file re-organisation:
  - Zero public-API changes — builder.rs's 59 call sites to detection::*
    continue to resolve unchanged via pub-use re-exports in mod.rs
  - No signature or behaviour changes
  - Tests split alongside the functions they cover
  - Largest new file is 347 LOC, well under the 500-LOC warning threshold

Verified via just preflight: clippy + arch-check + 6650-test octarine
suite all green.

Refs #193
