+++ crates/octarine/src/primitives/identifiers/government/detection/mexico.rs
@@ -1,0 +1,43 @@
+//! Mexico CURP detection
+
+use super::super::super::common::patterns;
+use super::super::super::types::{IdentifierMatch, IdentifierType};
+use super::helpers::{
+    MAX_IDENTIFIER_LENGTH, MAX_INPUT_LENGTH, deduplicate_matches, exceeds_safe_length,
+    get_full_match,
+};
+
+/// Check if a value matches Mexico CURP format
+#[must_use]
+pub fn is_mexico_curp(value: &str) -> bool {
+    if exceeds_safe_length(value, MAX_IDENTIFIER_LENGTH) {
+        return false;
+    }
+    patterns::mexico_curp::all()
+        .iter()
+        .any(|p| p.is_match(value))
+}
+
+/// Find all Mexico CURP patterns in text
+#[must_use]
+pub fn find_mexico_curps_in_text(text: &str) -> Vec<IdentifierMatch> {
+    if exceeds_safe_length(text, MAX_INPUT_LENGTH) {
+        return Vec::new();
+    }
+
+    let mut matches = Vec::new();
+
+    for pattern in patterns::mexico_curp::all() {
+        for capture in pattern.captures_iter(text) {
+            let full_match = get_full_match(&capture);
+            matches.push(IdentifierMatch::high_confidence(
+                full_match.start(),
+                full_match.end(),
+                full_match.as_str().to_string(),
+                IdentifierType::MexicoCurp,
+            ));
+        }
+    }
+
+    deduplicate_matches(matches)
+}
